+++ src/Adapters/Movies.Persistance.Postgres/Configurations/MovieConfiguration.cs
@@ -1,0 +1,33 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+using Movies.Domain;
+
+namespace Movies.Persistance.Postgres.Configurations;
+
+internal sealed class MovieConfiguration : IEntityTypeConfiguration<Movie>
+{
+    public void Configure(EntityTypeBuilder<Movie> builder)
+    {
+        builder.ToTable("movies");
+
+        builder.HasKey(m => m.Id);
+
+        builder.Property(m => m.Id)
+            .HasColumnName("id")
+            .ValueGeneratedOnAdd();
+
+        builder.Property(m => m.Title)
+            .HasColumnName("title")
+            .HasMaxLength(500)
+            .IsRequired();
+
+        builder.Property(m => m.YearOfRelease)
+            .HasColumnName("year_of_release")
+            .IsRequired();
+
+        builder.Property(m => m.Description)
+            .HasColumnName("description")
+            .HasMaxLength(2000);
+    }
+}