+++ src/Adapters/Movies.Persistance.Postgres/MoviesDbContext.cs
@@ -1,0 +1,15 @@
+using Microsoft.EntityFrameworkCore;
+
+using Movies.Domain;
+
+namespace Movies.Persistance.Postgres;
+
+public class MoviesDbContext(DbContextOptions<MoviesDbContext> options) : DbContext(options)
+{
+    public DbSet<Movie> Movies => Set<Movie>();
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.ApplyConfigurationsFromAssembly(typeof(MoviesDbContext).Assembly);
+    }
+}