+++ src/Applications/Movies.WebService/Middleware/ExceptionHandlingMiddleware.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
 using System.Text.Json;
 
-namespace Movies.WebService.Middleware;
+namespace Movies.WebService;
 
 public class ExceptionHandlingMiddleware(RequestDelegate next, ILogger<ExceptionHandlingMiddleware> logger)
 {